Also fixed some errors that cropped up during the merge.

Conflicts:
	scripting/methods.cpp

+++ lua_helpers.h
@@ -3,7 +3,7 @@
 
 void MakeTableItem (lua_State *L, const char * name, const CString & str);
 void MakeTableItem (lua_State *L, const char * name, const string & str);
-void MakeTableItem (lua_State *L, const char * name, const int n);
+void MakeTableItem (lua_State *L, const char * name, const double n);
 void MakeTableItem (lua_State *L, const char * name, const COleDateTime d);
 void MakeTableItemBool (lua_State *L, const char * name, const bool b);
 const bool optboolean (lua_State *L, const int narg, const int def);
